+++ src/Plugin/ModelMethodGenerator/Timestamp.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Drupal\wmscaffold\Plugin\ModelMethodGenerator;
+
+/**
+ * @ModelMethodGenerator(
+ *     id = "timestamp",
+ *     provider = "core",
+ * )
+ */
+class Timestamp extends StringType
+{
+}